requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
groq==0.5.0
python-dotenv==1.0.1
pymongo==4.6.1
dnspython==2.6.1
PyJWT==2.8.0
motor==3.3.2